move evaluateExpr etc. from ses to xs-npm/@agoric/evaluate
